+++ mlir/test/Dialect/MIOpen/Tuning/db_load_fail.mlir
@@ -1,4 +1,4 @@
-// RUN: miopen-gen -p %s | mlir-miopen-driver -miopen-affix-params -miopen-lowering -debug 2>&1 | FileCheck %s
+// RUN: miopen-gen -p %s | mlir-miopen-driver -miopen-affix-params -miopen-conv-to-gemm -debug 2>&1 | FileCheck %s
 
 // CHECK: Successfully opened connection to PerfDb
 // Perfdb does not have an entry for default configurations
