+++ mlir/test/Dialect/MIOpen/lowering_in_bounds_ops.mlir
@@ -1,4 +1,4 @@
-// RUN: miopen-opt -miopen-expand-shorthand %s | FileCheck %s
+// RUN: miopen-opt -miopen-sugar-to-loops %s | FileCheck %s
 
 module {
 // CHECK-LABEL: func @in_bounds_load_scalar
