+++ mlir/test/fusion/xdlops/fused_ir_tosa_conv7x7.mlir
@@ -1,4 +1,4 @@
-//RUN: mlir-miopen-driver -host-pipeline highlevel --miopen-affix-params --miopen-lowering --miopen-lowering-step2 --canonicalize --miopen-linalg-align %s | FileCheck %s
+//RUN: mlir-miopen-driver -host-pipeline highlevel --miopen-affix-params --miopen-conv-to-gemm --miopen-gridwise-gemm-to-blockwise --miopen-linalg-align %s | FileCheck %s
 module {
   func @main(%arg0: tensor<1x64x112x112xf32>, %arg1: tensor<1x3x224x224xf32>, %arg2: tensor<64x3x7x7xf32>) -> tensor<1x64x112x112xf32> attributes {kernel} {
     %cst = arith.constant dense<[0, 2, 3, 1]> : tensor<4xi64>
