+++ mlir/test/mlir-miopen-driver/populate_bwd_multi_kernels.mlir
@@ -1,12 +1,12 @@
-// RUN: miopen-gen  -fil_layout=gkcyx -in_layout=ngchw -out_layout=ngkhw -batchsize=32 -in_channels=32 -out_channels=32 -in_h=14 -in_w=14 -fil_h=3 -fil_w=3 --dilation_h=1 --dilation_w=1 --padding_h=1 --padding_w=1 --conv_stride_h=2 --conv_stride_w=2 --groupsize=1 --operation=conv2d_bwd_data | miopen-opt -miopen-affix-params -miopen-lowering | FileCheck %s --check-prefix=STRIDE2
+// RUN: miopen-gen  -fil_layout=gkcyx -in_layout=ngchw -out_layout=ngkhw -batchsize=32 -in_channels=32 -out_channels=32 -in_h=14 -in_w=14 -fil_h=3 -fil_w=3 --dilation_h=1 --dilation_w=1 --padding_h=1 --padding_w=1 --conv_stride_h=2 --conv_stride_w=2 --groupsize=1 --operation=conv2d_bwd_data | miopen-opt -miopen-affix-params -miopen-conv-to-gemm | FileCheck %s --check-prefix=STRIDE2
 
-// RUN: miopen-gen -fil_layout=gkyxc -in_layout=nhwgc -out_layout=nhwgk -batchsize=32 -in_channels=32 -out_channels=32 -in_h=14 -in_w=14 -fil_h=3 -fil_w=3 --dilation_h=1 --dilation_w=1 --padding_h=1 --padding_w=1 --conv_stride_h=2 --conv_stride_w=2 --groupsize=1  --operation=conv2d_bwd_data | miopen-opt -miopen-affix-params -miopen-lowering | FileCheck %s --check-prefix=STRIDE2_GKYXC
+// RUN: miopen-gen -fil_layout=gkyxc -in_layout=nhwgc -out_layout=nhwgk -batchsize=32 -in_channels=32 -out_channels=32 -in_h=14 -in_w=14 -fil_h=3 -fil_w=3 --dilation_h=1 --dilation_w=1 --padding_h=1 --padding_w=1 --conv_stride_h=2 --conv_stride_w=2 --groupsize=1  --operation=conv2d_bwd_data | miopen-opt -miopen-affix-params -miopen-conv-to-gemm | FileCheck %s --check-prefix=STRIDE2_GKYXC
 
 // This config requires a zero initialization utility kernel.
 // Check at the top-level there is a utility kernel with gemm_id=-1.
 // RUN: miopen-gen  -fil_layout=gkcyx -in_layout=ngchw -out_layout=ngkhw -batchsize=32 -in_channels=32 -out_channels=32 -in_h=14 -in_w=14 -fil_h=1 -fil_w=1 --dilation_h=1 --dilation_w=1 --padding_h=1 --padding_w=1 --conv_stride_h=2 --conv_stride_w=2 --groupsize=1  --operation=conv2d_bwd_data | FileCheck %s --check-prefix=STRIDE2_1x1_TOP_LEVEL
 // Check after -miopen-lowering, only gridwise_gemm with gemm_id=0 exists.
-// RUN: miopen-gen  -fil_layout=gkcyx -in_layout=ngchw -out_layout=ngkhw -batchsize=32 -in_channels=32 -out_channels=32 -in_h=14 -in_w=14 -fil_h=1 -fil_w=1 --dilation_h=1 --dilation_w=1 --padding_h=1 --padding_w=1 --conv_stride_h=2 --conv_stride_w=2 --groupsize=1  --operation=conv2d_bwd_data | miopen-opt -miopen-affix-params -miopen-lowering | FileCheck %s --check-prefix=STRIDE2_1x1_LOWERING
+// RUN: miopen-gen  -fil_layout=gkcyx -in_layout=ngchw -out_layout=ngkhw -batchsize=32 -in_channels=32 -out_channels=32 -in_h=14 -in_w=14 -fil_h=1 -fil_w=1 --dilation_h=1 --dilation_w=1 --padding_h=1 --padding_w=1 --conv_stride_h=2 --conv_stride_w=2 --groupsize=1  --operation=conv2d_bwd_data | miopen-opt -miopen-affix-params -miopen-conv-to-gemm | FileCheck %s --check-prefix=STRIDE2_1x1_LOWERING
 
 // STRIDE2: {{miopen.gridwise_gemm.*gemm_id = 0 : i32.*matrix_a_source_data_per_read = 1 : i32, matrix_a_source_vector_read_dim = 2 : i32.*matrix_b_source_data_per_read = 1 : i32, matrix_b_source_vector_read_dim = 2 : i32.*}}
 // STRIDE2: {{miopen.gridwise_gemm.*gemm_id = 1 : i32.*matrix_a_source_data_per_read = 1 : i32, matrix_a_source_vector_read_dim = 2 : i32.*matrix_b_source_data_per_read = 1 : i32, matrix_b_source_vector_read_dim = 2 : i32.*}}
