+++ mlir/test/mlir-miopen-driver/tuning_xdlops.mlir
@@ -1,8 +1,8 @@
 // Check the naming of tuning parameters for xdlops and matrix c vectorization values
 
-// RUN: miopen-gen -p -x2 | mlir-miopen-driver -miopen-affix-params -miopen-lowering | FileCheck %s --check-prefix=STEP1
-// RUN: miopen-gen -p -x2 | mlir-miopen-driver -miopen-affix-params -miopen-lowering -miopen-lowering-step2 | FileCheck %s --check-prefix=STEP2
-// RUN: miopen-gen -p --fil_layout=kyxc --in_layout=nhwc --out_layout=nhwk -x2 | mlir-miopen-driver -miopen-affix-params -miopen-lowering | FileCheck %s --check-prefix=NHWC
+// RUN: miopen-gen -p -x2 | mlir-miopen-driver -miopen-affix-params -miopen-conv-to-gemm | FileCheck %s --check-prefix=STEP1
+// RUN: miopen-gen -p -x2 | mlir-miopen-driver -miopen-affix-params -miopen-conv-to-gemm -miopen-gridwise-gemm-to-blockwise | FileCheck %s --check-prefix=STEP2
+// RUN: miopen-gen -p --fil_layout=kyxc --in_layout=nhwc --out_layout=nhwk -x2 | mlir-miopen-driver -miopen-affix-params -miopen-conv-to-gemm | FileCheck %s --check-prefix=NHWC
 
 // STEP1: m_per_wave
 // STEP1: matrix_c_data_per_copy = 4
