+++ mlir/utils/generateInWarpTransposeTest.py
@@ -18,9 +18,9 @@ def testHeader() -> str:
 // Generated by mlir/utils/generateInWarpTransposeTest.py
 
 // RUN: miopen-opt \\
-// RUN: --miopen-lowering-step4 --miopen-expand-shorthand --miopen-loops-to-cf %s \\
+// RUN: --miopen-sugar-to-loops --miopen-loops-to-cf --convert-miopen-to-gpu %s \\
 // RUN:  | mlir-rocm-runner \\
-// RUN: --shared-libs=%conv_validation_wrapper_library_dir/libconv-validation-wrappers%shlibext,%linalg_test_lib_dir/libmlir_runner_utils%shlibext \\
+// RUN: --shared-libs=%linalg_test_lib_dir/libmlir_rocm_runtime%shlibext,%conv_validation_wrapper_library_dir/libconv-validation-wrappers%shlibext,%linalg_test_lib_dir/libmlir_runner_utils%shlibext \\
 // RUN: --entry-point-result=i32 | FileCheck %s
 
 // CHECK: {{^}}0{{$}}
@@ -30,9 +30,8 @@ def testHeader() -> str:
 def gpuCode(spec: TestSpec) -> str:
     """Generates the gpu-side code for calling a particular inWarpTranspose test"""
     ret = [f"""
-gpu.module @{spec.name} {{
-gpu.func @{spec.name}_kern(%arg0: memref<1024xi32>, %arg1: memref<1024xi32>) -> () kernel
-attributes {{block_size = 64 : i32, grid_size = 64 : i32}} {{
+func.func @{spec.name}_kern(%arg0: memref<1024xi32>, %arg1: memref<1024xi32>) -> ()
+attributes {{block_size = 64 : i32, grid_size = 64 : i32, kernel = 0 : i32}} {{
   %cst0 = arith.constant 0 : index
   %cst0_i32 = arith.constant 0 : i32
   %cst1 = arith.constant 1 : index
@@ -63,8 +62,7 @@ gpu.barrier""")
 memref.store %e_{i}, %arg1[{store_idx}] : memref<1024xi32>
 %store_idx_{i+1} = arith.addi {store_idx}, %cst64 : index""")
     ret.append("""
-gpu.return
-}
+func.return
 }""")
     return '\n'.join(ret)
 
@@ -86,17 +84,7 @@ def hostCode(spec: TestSpec) -> str:
     goal = swizzleGoal(spec)
 
     ret = f"""
-func @host_{spec.name}_run(%arg0: memref<1024xi32>, %arg1: memref<1024xi32>) {{
-    %cst64 = arith.constant 64 : index
-    %cst = arith.constant 1 : index
-    gpu.launch_func @{spec.name}::@{spec.name}_kern
-        blocks in (%cst, %cst, %cst)
-        threads in (%cst64, %cst, %cst)
-        args(%arg0 : memref<1024xi32>, %arg1 : memref<1024xi32>)
-    return
-}}
-
-func @host_{spec.name}() -> i1 {{
+func.func @host_{spec.name}() -> i1 {{
     %init = arith.constant dense<{init}> : tensor<1024xi32>
     %goal = arith.constant dense<{goal}> :  tensor<1024xi32>
     %cst1_i32 = arith.constant 1 : i32
@@ -120,7 +108,7 @@ func @host_{spec.name}() -> i1 {{
     call @mgpuMemSetInt32(%res_gpu_dyn, %cst_deadbeef_i32) : (memref<?xi32>, i32) -> ()
     %res_gpu = memref.cast %res_gpu_dyn : memref<?xi32> to memref<1024xi32>
 
-    call @host_{spec.name}_run(%arg_gpu, %res_gpu) : (memref<1024xi32>, memref<1024xi32>) -> ()
+    call @{spec.name}_kern(%arg_gpu, %res_gpu) : (memref<1024xi32>, memref<1024xi32>) -> ()
 
     call @mgpuMemCopyInt32(%res_gpu_dyn, %res_dyn, %cst2_i32) : (memref<?xi32>, memref<?xi32>, i32) -> ()
 
@@ -140,7 +128,7 @@ func @host_{spec.name}() -> i1 {{
     }}
     call @mgpuMemDeallocInt32(%arg_gpu_dyn) : (memref<?xi32>) -> ()
     call @mgpuMemDeallocInt32(%res_gpu_dyn) : (memref<?xi32>) -> ()
-    return %ret : i1
+    func.return %ret : i1
 }}
 """
     return ret
@@ -148,11 +136,11 @@ func @host_{spec.name}() -> i1 {{
 def genTestProgram(specs: List[TestSpec]):
     ret: List[str] = [f"""
 module attributes {{gpu.container_module}} {{
-    func private @mgpuMemAllocInt32(%ptr : memref<?xi32>) -> (memref<?xi32>)
-    func private @mgpuMemDeallocInt32(%ptr : memref<?xi32>)
-    func private @mgpuMemSetInt32(%ptr : memref<?xi32>, %value: i32)
-    func private @mgpuMemCopyInt32(%src : memref<?xi32>, %dest : memref<?xi32>, %dir: i32)
-    func private @print_memref_i32(memref<*xi32>)
+    func.func private @mgpuMemAllocInt32(%ptr : memref<?xi32>) -> (memref<?xi32>)
+    func.func private @mgpuMemDeallocInt32(%ptr : memref<?xi32>)
+    func.func private @mgpuMemSetInt32(%ptr : memref<?xi32>, %value: i32)
+    func.func private @mgpuMemCopyInt32(%src : memref<?xi32>, %dest : memref<?xi32>, %dir: i32)
+    func.func private @print_memref_i32(memref<*xi32>)
 """]
     for spec in specs:
         ret.append(gpuCode(spec))
@@ -188,7 +176,8 @@ def lowerTestCode(testCode: str) -> str:
 
     result = subprocess.run(
         [optimizer, "--convert-linalg-to-affine-loops", "--lower-affine",
-        "--arith-bufferize", "--finalizing-bufferize", "-o", "-", "-"],
+        "--arith-bufferize", "--finalizing-bufferize", "--canonicalize",
+        "-o", "-", "-"],
         encoding="utf-8", input=testCode, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     if result.returncode != 0 or len(result.stderr) != 0:
         print("Couldn't lower successfully, miopen-opt returned:",
